Module ic::inet

Re-exports

pub use super::*;

Structs

Cursor

INet

INode

Constants

ANN

CON

DUP

ERA

FIX

REF

ROOT

TAG

TAG_MASK

Functions

addr

compare

enter

equal

from_string_inodes

get

inet_to_inodes

inodes_to_inet

kind

link

new_inet

new_node

normal

parse_inodes

port

reduce

rewrite

Rewrites an active pair. Returns None if rewrite happened, Some(next) otherwise. This is the case when y is REF node and was substituted by its body, and the body’s entry node doesn’t form an active pair with x, e.g. def A = (B C) The caller (reduce) can then use the returned port to continue advancing.

show_inodes

slot

to_string_inodes

Type Definitions

Chr

NodeId

NodeKind

Port

SlotId

Str